add fixture for creating profile
